# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ORION — Orbital Real-time Intelligence & Operations Network API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SpaceObject {
  /** NORAD catalog number */
  noradId: string;
  name: string;
  /** One of: stations, starlink, gps, debris, active, military, weather */
  category: string;
  /** Current geodetic latitude in degrees */
  lat: number;
  /** Current geodetic longitude in degrees */
  lon: number;
  /** Current altitude in kilometers */
  alt: number;
  /** Orbital velocity in km/s */
  velocity: number;
  /** Orbital inclination in degrees */
  inclination: number;
  /** Orbital period in minutes */
  period: number;
  /**
     * Radar cross-section size class (SMALL, MEDIUM, LARGE)
     * @nullable
     */
  rcs: string | null;
  /** @nullable */
  launchYear: number | null;
  isDebris: boolean;
  isActive: boolean;
}

export interface SpaceObjectDetail {
  noradId: string;
  name: string;
  category: string;
  lat: number;
  lon: number;
  alt: number;
  velocity: number;
  inclination: number;
  period: number;
  /** @nullable */
  rcs: string | null;
  /** @nullable */
  launchYear: number | null;
  isDebris: boolean;
  isActive: boolean;
  /** TLE line 1 (raw) */
  tleLine1: string;
  /** TLE line 2 (raw) */
  tleLine2: string;
  /** Apogee altitude in km */
  apogee: number;
  /** Perigee altitude in km */
  perigee: number;
  eccentricity: number;
  /** TLE epoch as ISO timestamp */
  epoch: string;
  /** Right ascension of ascending node in degrees */
  rightAscension: number;
  /** Argument of perigee in degrees */
  argOfPerigee: number;
  /** Mean anomaly in degrees */
  meanAnomaly: number;
  /** BSTAR drag term */
  dragTerm: number;
}

export interface ObjectCategory {
  id: string;
  label: string;
  count: number;
  /** Hex color for this category */
  color: string;
}

export interface FleetStats {
  totalTracked: number;
  activeCount: number;
  debrisCount: number;
  stationsCount: number;
  starlinkCount: number;
  gpsCount: number;
  weatherCount: number;
  militaryCount: number;
  /** NOMINAL, ELEVATED, or CRITICAL */
  threatLevel: string;
  /** ISO timestamp of last TLE cache refresh */
  lastUpdated: string;
  avgAltitude: number;
  /**
     * Name of nearest conjunction candidate
     * @nullable
     */
  highestThreatObject: string | null;
}

export interface ConjunctionEvent {
  id: string;
  primaryName: string;
  primaryNoradId: string;
  secondaryName: string;
  secondaryNoradId: string;
  /** Predicted minimum range in km */
  minRange: number;
  /** Collision probability (0-1) */
  probability: number;
  /** ISO timestamp */
  timeOfClosestApproach: string;
  /** Relative velocity in km/s */
  relativeVelocity: number;
  /** LOW, MEDIUM, HIGH, CRITICAL */
  threatLevel: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListSatellitesParams = {
/**
 * Filter by object category
 */
category?: string;
/**
 * Max results to return
 */
limit?: number;
/**
 * Search by name
 */
search?: string;
};