# lib/api-zod/src/generated/types/conjunctionEvent.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ORION — Orbital Real-time Intelligence & Operations Network API
 * OpenAPI spec version: 0.1.0
 */

export interface ConjunctionEvent {
  id: string;
  primaryName: string;
  primaryNoradId: string;
  secondaryName: string;
  secondaryNoradId: string;
  /** Predicted minimum range in km */
  minRange: number;
  /** Collision probability (0-1) */
  probability: number;
  /** ISO timestamp */
  timeOfClosestApproach: string;
  /** Relative velocity in km/s */
  relativeVelocity: number;
  /** LOW, MEDIUM, HIGH, CRITICAL */
  threatLevel: string;
}