# lib/api-zod/src/generated/types/fleetStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ORION — Orbital Real-time Intelligence & Operations Network API
 * OpenAPI spec version: 0.1.0
 */

export interface FleetStats {
  totalTracked: number;
  activeCount: number;
  debrisCount: number;
  stationsCount: number;
  starlinkCount: number;
  gpsCount: number;
  weatherCount: number;
  militaryCount: number;
  /** NOMINAL, ELEVATED, or CRITICAL */
  threatLevel: string;
  /** ISO timestamp of last TLE cache refresh */
  lastUpdated: string;
  avgAltitude: number;
  /**
     * Name of nearest conjunction candidate
     * @nullable
     */
  highestThreatObject: string | null;
}