# lib/api-zod/src/generated/types/objectCategory.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ORION — Orbital Real-time Intelligence & Operations Network API
 * OpenAPI spec version: 0.1.0
 */

export interface ObjectCategory {
  id: string;
  label: string;
  count: number;
  /** Hex color for this category */
  color: string;
}