# lib/api-zod/src/generated/types/spaceObject.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ORION — Orbital Real-time Intelligence & Operations Network API
 * OpenAPI spec version: 0.1.0
 */

export interface SpaceObject {
  /** NORAD catalog number */
  noradId: string;
  name: string;
  /** One of: stations, starlink, gps, debris, active, military, weather */
  category: string;
  /** Current geodetic latitude in degrees */
  lat: number;
  /** Current geodetic longitude in degrees */
  lon: number;
  /** Current altitude in kilometers */
  alt: number;
  /** Orbital velocity in km/s */
  velocity: number;
  /** Orbital inclination in degrees */
  inclination: number;
  /** Orbital period in minutes */
  period: number;
  /**
     * Radar cross-section size class (SMALL, MEDIUM, LARGE)
     * @nullable
     */
  rcs: string | null;
  /** @nullable */
  launchYear: number | null;
  isDebris: boolean;
  isActive: boolean;
}