# lib/api-zod/src/generated/types/spaceObjectDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ORION — Orbital Real-time Intelligence & Operations Network API
 * OpenAPI spec version: 0.1.0
 */

export interface SpaceObjectDetail {
  noradId: string;
  name: string;
  category: string;
  lat: number;
  lon: number;
  alt: number;
  velocity: number;
  inclination: number;
  period: number;
  /** @nullable */
  rcs: string | null;
  /** @nullable */
  launchYear: number | null;
  isDebris: boolean;
  isActive: boolean;
  /** TLE line 1 (raw) */
  tleLine1: string;
  /** TLE line 2 (raw) */
  tleLine2: string;
  /** Apogee altitude in km */
  apogee: number;
  /** Perigee altitude in km */
  perigee: number;
  eccentricity: number;
  /** TLE epoch as ISO timestamp */
  epoch: string;
  /** Right ascension of ascending node in degrees */
  rightAscension: number;
  /** Argument of perigee in degrees */
  argOfPerigee: number;
  /** Mean anomaly in degrees */
  meanAnomaly: number;
  /** BSTAR drag term */
  dragTerm: number;
}